feat: decimal token-proportional credits (10x markup), 회→c unit change
- tier_config: MODEL_CREDIT_RATES, TIER_ESTIMATES, calculate_credit_cost()
- ai_client: return (result, total_tokens) from structured AI calls
- usage_service: float amounts, adjust_credit() for worker reconciliation
- diagram/retry: token-proportional charging with balance gate pattern
- models/schemas: UsageRecord.consumed Integer→Float, float types
- migration: usage_records.consumed column type change
- frontend: credit unit 회→c, plan allocations 5/60/200/700c, decimal display

+++ backend/app/schemas/billing.py
@@ -14,8 +14,8 @@ class _CamelModel(BaseModel):
 
 class UsageWindowSchema(_CamelModel):
     resource_type: str
-    consumed: int
-    limit: int
+    consumed: float
+    limit: float
     window_starts_at: datetime
     window_ends_at: datetime
     source: str
@@ -61,6 +61,6 @@ class ManageUrlsResponse(_CamelModel):
 class LimitExceededError(_CamelModel):
     detail: str
     limit_type: str
-    current_usage: int
-    limit: int
+    current_usage: float
+    limit: float
     upgrade_url: str = "/pricing"
